Split search route into separate module
