+++ app/api/routes/__init__.py
@@ -3,3 +3,4 @@ from app.api.routes import resources
 from app.api.routes import resource_voting
 from app.api.routes import categories
 from app.api.routes import languages
+from app.api.routes import searching
